Fix for bug #16249
- Making removal of a menu item tied to a content item not remove that content item, instead the OnMainMenu bit is flipped
- Changing the text of the remove button to be either "Remove" (for items tied to content) or "Delete" (for standalone items)

--HG--
branch : dev

+++ src/Orchard.Web/Core/Navigation/Controllers/AdminController.cs
@@ -108,8 +108,12 @@ namespace Orchard.Core.Navigation.Controllers {
 
             MenuPart menuPart = _menuService.Get(id);
 
-            if (menuPart != null)
-                _menuService.Delete(menuPart);
+            if (menuPart != null) {
+                if (menuPart.Is<MenuItem>())
+                    _menuService.Delete(menuPart);
+                else
+                    menuPart.OnMainMenu = false;
+            }
 
             return RedirectToAction("Index");
         }
